requirements.txt
# === Web Framework ===
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# === MetaAPI Trading SDK ===
metaapi-cloud-sdk==28.0.3

# === Data & Analysis ===
pandas==2.2.2
pandas-ta==0.3.14b1
numpy==1.26.4

# === Machine Learning ===
xgboost==2.1.0
lightgbm==4.5.0
scikit-learn==1.5.1
joblib==1.4.2

# === Database ===
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2

# === Cache & Task Queue ===
redis==5.0.8
celery==5.4.0

# === HTTP & WebSocket ===
httpx==0.27.0
websockets==12.0

# === NLP / Sentiment ===
vaderSentiment==3.3.2
google-generativeai==0.8.0

# === Configuration ===
python-dotenv==1.0.1
pydantic-settings==2.4.0

# === Utilities ===
python-dateutil==2.9.0
pytz==2024.1

# === Testing ===
pytest==8.3.2
pytest-asyncio==0.23.8
